+++ src/setenv.rs
@@ -1,5 +1,6 @@
 use std::env::set_var;
 use std::collections::HashMap;
+// This is how to import other files from rush when the current source file is imported in "rush.rs"
 use crate::helpful::*;
 
 pub fn setenv(args:&[String], index:usize, returns:&mut HashMap<usize, CommandStatus>) {