+++ changes.md
@@ -36,10 +36,15 @@
 
 - RUSH: Support for keywords starting with "$"
 
-### Waiting features:
+### Added on 09.04.2024
 
 - RUSH: Avoid word splitting when words are enclosed in "" or ''
-- CMP: Ability to compare outputs from any variable or command
+- RUSH: Known bug - Words in quotes are always added to the command as the last
+- RUSH: Known bug - Shell is not testing if quotes are properly ended or not
+- CMP: Ability to compare with command's stdout, stderr or exit code. It'll also work with variables because the shell itself is responsible for replacing text starting with a dollar sign with a variable contents.
+
+### Waiting features:
+
 - MATH: Do the math stuff
 - WHILE: LOOP, MATCH, FOR: More logic operations
 - BREAK: Quit from the loop